refactor(table): update dto name files
